third_party/blink/web_tests/virtual/computed-width-values-depend-on-style-values/external/wpt/css/css-multicol/parsing/column-rule-width-computed-expected.txt

This is a testharness.js-based test.
[FAIL] column-rule-width is as specified when column-rule-style is none or hidden
  assert_equals: expected "10px" but got "0px"
Harness: the test ran to completion.
